ignore text sizes for jumping in the tree
getting the exact width of a rendered string is expensive. the only real
way to do this with pango is to actually compute the glyph layout. doing
it for every label in a fully unfolded SSU-Ref (500k sequences) takes
seconds to minutes.

if there is no major drawbacks, maybe a backport to trunk would make
sense, also. it saves calculating the nds strings for every leaf just 
to find the jump position in the tree.