fixes PT server never updated/built if MEMORY_TEST evalutes to true
This was originally broken in r9242 by removing the "build_map" entry from 
struct build_step[], which changed the meaning of testing for index 1 
("s == 1") within the loop. To avoid further similar issues, the loop
is unrolled by this patch as this seems to be an example of DRY not 
always being the safest option.

(the patch made it into branches/ptsfix in r9447 and into trunk in r11060)